Implementing the Onion Architecture to the console application
