+++ Mastermind.Application/Randomizer/IRandomizer.cs
@@ -1,0 +1,10 @@
+using Mastermind.Domain;
+
+namespace Mastermind.Application.Randomizer;
+
+public interface IRandomizer
+{
+    public Colour[] GetRandomColours(int numberOfColoursToSelect);
+
+    IEnumerable<Clue> GetShuffledArray(IEnumerable<Clue> orderedArray);
+}
