+++ Mastermind.Domain/Clue.cs
@@ -1,0 +1,7 @@
+namespace Mastermind.Domain;
+
+public enum Clue
+{
+    Black,
+    White
+}
