+++ Mastermind.Domain/GameStatus.cs
@@ -1,0 +1,9 @@
+namespace Mastermind.Domain;
+
+public enum GameStatus
+{
+    Playing,
+    Won,
+    Lost,
+    Quit
+}
