+++ Mastermind.Presentation/InputOutput/ColourSquares.cs
@@ -1,0 +1,15 @@
+namespace Mastermind.Presentation.InputOutput;
+
+public static class ColourSquares
+{
+    public const string Red   = "🟥";
+    public const string Blue  = "🟦";
+    public const string Green = "🟩";
+    public const string Orange = "🟧";
+    public const string Purple = "🟪";
+    public const string Yellow = "🟨";
+        
+    public const string Black = "⬛";
+    public const string White = "⬜";
+
+}
