+++ Mastermind.Presentation/InputOutput/IInputOutput.cs
