+++ Mastermind.Presentation/InputOutput/IOutput.cs
