+++ Mastermind.Presentation/Mastermind.Presentation.csproj
@@ -1,0 +1,15 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <OutputType>Exe</OutputType>
+        <TargetFramework>net6.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\Mastermind.Application\Mastermind.Application.csproj" />
+      <ProjectReference Include="..\Mastermind.Domain\Mastermind.Domain.csproj" />
+    </ItemGroup>
+
+</Project>
