+++ Mastermind.sln
@@ -1,8 +1,12 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Mastermind", "Mastermind\Mastermind.csproj", "{B472DAA3-FEF2-4CBF-AD4B-BFA2366EBA4F}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MastermindTests", "MastermindTests\MastermindTests.csproj", "{37738FBF-8DDB-4268-BBE2-123A2C27173A}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Mastermind.Domain", "Mastermind.Domain\Mastermind.Domain.csproj", "{5D6DA2F1-3296-4DC7-AFD7-98A79FEF404B}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Mastermind.Application", "Mastermind.Application\Mastermind.Application.csproj", "{C521CBD6-2751-4B5B-9FA1-98F6B83DD4DC}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Mastermind.Presentation", "Mastermind.Presentation\Mastermind.Presentation.csproj", "{0A4821AC-FC2E-4B26-B99F-5AD3C7C58E1E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -10,13 +14,21 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{B472DAA3-FEF2-4CBF-AD4B-BFA2366EBA4F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B472DAA3-FEF2-4CBF-AD4B-BFA2366EBA4F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B472DAA3-FEF2-4CBF-AD4B-BFA2366EBA4F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B472DAA3-FEF2-4CBF-AD4B-BFA2366EBA4F}.Release|Any CPU.Build.0 = Release|Any CPU
 		{37738FBF-8DDB-4268-BBE2-123A2C27173A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{37738FBF-8DDB-4268-BBE2-123A2C27173A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{37738FBF-8DDB-4268-BBE2-123A2C27173A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{37738FBF-8DDB-4268-BBE2-123A2C27173A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5D6DA2F1-3296-4DC7-AFD7-98A79FEF404B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5D6DA2F1-3296-4DC7-AFD7-98A79FEF404B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5D6DA2F1-3296-4DC7-AFD7-98A79FEF404B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5D6DA2F1-3296-4DC7-AFD7-98A79FEF404B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C521CBD6-2751-4B5B-9FA1-98F6B83DD4DC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C521CBD6-2751-4B5B-9FA1-98F6B83DD4DC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C521CBD6-2751-4B5B-9FA1-98F6B83DD4DC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C521CBD6-2751-4B5B-9FA1-98F6B83DD4DC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0A4821AC-FC2E-4B26-B99F-5AD3C7C58E1E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0A4821AC-FC2E-4B26-B99F-5AD3C7C58E1E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0A4821AC-FC2E-4B26-B99F-5AD3C7C58E1E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0A4821AC-FC2E-4B26-B99F-5AD3C7C58E1E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
