--- Mastermind/Domain/Models/Clue.cs
@@ -1,8 +1,0 @@
-namespace Mastermind.Domain.Models
-{
-    public enum Clue
-    {
-        Black,
-        White
-    }
-}
