--- Mastermind/Domain/Models/GameStatus.cs
@@ -1,10 +1,0 @@
-namespace Mastermind.Domain.Models
-{
-    public enum GameStatus
-    {
-        Playing,
-        Won,
-        Lost,
-        Quit
-    }
-}
