--- Mastermind/Randomizer/IRandomizer.cs
@@ -1,11 +1,0 @@
-using Mastermind.Domain.Models;
-
-namespace Mastermind.Randomizer
-{
-    public interface IRandomizer
-    {
-        public Colour[] GetRandomColours(int numberOfColoursToSelect);
-
-        IEnumerable<Clue> GetShuffledArray(IEnumerable<Clue> orderedArray);
-    }
-}
