+++ MastermindTests/MastermindTests.csproj
@@ -22,6 +22,7 @@
     </ItemGroup>
 
     <ItemGroup>
+      <ProjectReference Include="..\Mastermind.Presentation\Mastermind.Presentation.csproj" />
       <ProjectReference Include="..\Mastermind\Mastermind.csproj" />
     </ItemGroup>
